ThemeUtils tests
#1504
